+++ src/App.tsx
@@ -1,4 +1,4 @@
-import MainPage from "./pages/HomePage";
+import MainPage from "./pages/MainPage";
 import { BrowserRouter } from "react-router-dom";
 
 function App() {
