+++ src/pages/MainPageStyles.ts
